requirements.txt
pandas==2.2.1
requests
beautifulsoup4==4.12.3
snowflake-connector-python>=3.12.1
apache-airflow==2.10.1